refactored tree changes
